+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-- Bug Fix: [Add `__private__` field to EnumValueDefinition](https://github.com/absinthe-graphql/absinthe/pull/1148)
 - Feature: [Convert SDL Language.\* structs to SDL notation](https://github.com/absinthe-graphql/absinthe/pull/1160)
+- Bug Fix: [Add `__private__` field to EnumValueDefinition](https://github.com/absinthe-graphql/absinthe/pull/1148)
+- Bug Fix: [Fix bug in Schema.__absinthe_types__(:all) for Persistent Term](https://github.com/absinthe-graphql/absinthe/pull/1161)
 
 ## 1.7.0